Startup damping + predicted positions
Add startup damping on the pressure multiplier and acceleration. This allows particles close together or close to the edge to ease away from each other without gaining massive EK. We also compute the acceleration from the next predicted position instead of from the current position of the particle. I'm not sure why this works but it makes the simulation much less chaotic and prevents runaway EK.